requirements.txt
numpy
pandas
torch
tensorboard
torchsummary
pyyaml
tqdm

-e .